Fixed the duplicated entry issue.
Now the 'unique-timestamps' feature is enabled by default. This means that a Virtual Sensor defined by a description file which does not set the 'unique-timesamps' will display WARNING messages and do not put duplicated data into DB.

If you want to allow your Virtual Sensor to put duplicated data into DB, add the following tag in the VS xml configuration file.

<unique-timestamps>false</unique-timestamps>


WARNING: At this time, the 'unique-timestamps' feature is enabled or disabled only at the very first time a Virtual Sensor is loaded into GSN. Enabling or disabling this feature on an already loaded VS could be done by manually modifying the 'timed' index associated to its table in the DB. Set the 'timed' index as UNIQUE if you want to enable the 'unique-timestamps' feature.
In case you need to enabled this feature, notice that you should also manually remove the duplicated data from the DB.