Incorporporated the old Chapter 5 - Writing a Wrapper into Appendix A.  As before, review by someone with a better understanding of the process would be appreciated.
I also note that the format of the new document supports only three levels of numbered headings (chapter, section and subsection with subsubsection producing an heading without a number) whereas the old doc went to four levels.  I thinks the clarity of some of this material would be enhanced by an additional level but don't really have the Latex skills to fix it.

Can anyone please help.