When I tried a scheduled stream exporter with a longer interval (86400000) to generate a daily output, I found that the database connection had timed after at 8 hours which is the default for MySql.
One options to overcome this would be to increase the timeout to 24 hours or to specifically open the connection for each write.  As the most likely use of a scheduled stream exporter would involve longer intervals I opted to specifically open the connection before writing and close it again after.